+++ samples/broker_start.py
@@ -1,8 +1,13 @@
 import asyncio
 import logging
 import os
+from pathlib import Path
 
 from amqtt.broker import Broker
+
+"""
+This sample shows how to run a broker without stacktraces on keyboard interrupt
+"""
 
 logger = logging.getLogger(__name__)
 
@@ -18,31 +23,50 @@ config = {
             "max_connections": 10,
         },
     },
-    "sys_interval": 10,
-    "auth": {
-        "allow-anonymous": True,
-        "password-file": os.path.join(
-            os.path.dirname(os.path.realpath(__file__)),
-            "passwd",
-        ),
-        "plugins": ["auth_file", "auth_anonymous"],
-    },
-    "topic-check": {"enabled": False},
+    "plugins": {
+        'amqtt.plugins.authentication.AnonymousAuthPlugin': { 'allow_anonymous': True},
+        'amqtt.plugins.authentication.FileAuthPlugin': {
+            'password_file': Path(__file__).parent / 'passwd',
+        },
+        'amqtt.plugins.sys.broker.BrokerSysPlugin': { "sys_interval": 10},
+    }
 }
 
-broker = Broker(config)
+async def main_loop():
+    broker = Broker(config)
+    try:
+        await broker.start()
+        while True:
+            await asyncio.sleep(1)
+    except asyncio.CancelledError:
+        await broker.shutdown()
 
+async def main():
+    t = asyncio.create_task(main_loop())
+    try:
+        await t
+    except asyncio.CancelledError:
+        pass
 
-async def test_coro() -> None:
-    await broker.start()
-    # await asyncio.sleep(5)
-    # await broker.shutdown()
+def __main__():
 
+    formatter = "[%(asctime)s] :: %(levelname)s :: %(name)s :: %(message)s"
+    logging.basicConfig(level=logging.INFO, format=formatter)
+
+    loop = asyncio.new_event_loop()
+    asyncio.set_event_loop(loop)
+
+    task = loop.create_task(main())
+
+    try:
+        loop.run_until_complete(task)
+    except KeyboardInterrupt:
+        logger.info("KeyboardInterrupt received. Stopping server...")
+        task.cancel()
+        loop.run_until_complete(task)  # Ensure task finishes cleanup
+    finally:
+        logger.info("Server stopped.")
+        loop.close()
 
 if __name__ == "__main__":
-    formatter = "[%(asctime)s] :: %(levelname)s :: %(name)s :: %(message)s"
-    # formatter = "%(asctime)s :: %(levelname)s :: %(message)s"
-    logging.basicConfig(level=logging.INFO, format=formatter)
-    asyncio.get_event_loop().run_until_complete(test_coro())
-    asyncio.get_event_loop().run_forever()
-
+    __main__()